+++ CHANGELOG.md
@@ -21,6 +21,9 @@ ingress.
   - `jobs listtasks`: `--jobid` allows scoping of list tasks to a specific job
 
 ### Changed
+- **Breaking Change:** Properties of `docker_registry` have changed
+significantly to support integration with the Azure Container Registry
+service. Please see the configuration doc for more information.
 - `files` data ingress no longer creates a directory where files to
 be uploaded exist. For example if uploading from a path `/a/b/c`, the
 directory `c` is no longer created at the destination. Instead all files